Begin replacing ints with floats in SmyLibGui

+++ src/main/java/fr/thesmyler/smylibgui/widgets/IWidget.java
@@ -9,12 +9,12 @@ public interface IWidget {
 	/**
 	 * @return the x position of this widget in the parent container
 	 */
-	public int getX();
+	public float getX();
 
 	/**
 	 * @return the y position of this widget in the parent container
 	 */
-	public int getY();
+	public float getY();
 
 	/**
 	 * @return the z position of this widget in the parent container
@@ -24,12 +24,12 @@ public interface IWidget {
 	/**
 	 * @return the width of this widget
 	 */
-	public int getWidth();
+	public float getWidth();
 
 	/**
 	 * @return the height of this widget
 	 */
-	public int getHeight();
+	public float getHeight();
 	
 	/**
 	 * This method should draw the widget on the screen
@@ -42,7 +42,7 @@ public interface IWidget {
 	 * @param focused indicates whether or not this widget has its parent's focus (it will get keystrokes and so on)
 	 * @param parent the parent Screen
 	 */
-	public void draw(int x, int y, int mouseX, int mouseY, boolean hovered, boolean focused, @Nullable Screen parent);
+	public void draw(float x, float y, float mouseX, float mouseY, boolean hovered, boolean focused, @Nullable Screen parent);
 	
 	/**
 	 * If this returns false, this widget will not be rendered and or notified of user actions
@@ -78,7 +78,7 @@ public interface IWidget {
 	 * 
 	 * @return a boolean indicating whether or not this event should propagate to widgets with lower priorities
 	 */
-	public default boolean onClick(int mouseX, int mouseY, int mouseButton, @Nullable Screen parent) {
+	public default boolean onClick(float mouseX, float mouseY, int mouseButton, @Nullable Screen parent) {
 		return true;
 	}
 	
@@ -97,7 +97,7 @@ public interface IWidget {
 	 * 
 	 * @return a boolean indicating whether or not this event should propagate to widgets with lower priorities
 	 */
-	public default boolean onParentClick(int mouseX, int mouseY, int mouseButton, @Nullable Screen parent) {
+	public default boolean onParentClick(float mouseX, float mouseY, int mouseButton, @Nullable Screen parent) {
 		return true;
 	}
 	
@@ -115,7 +115,7 @@ public interface IWidget {
 	 * 
 	 * @return a boolean indicating whether or not this event should propagate to widgets with lower priorities
 	 */
-	public default boolean onDoubleClick(int mouseX, int mouseY, int mouseButton, @Nullable Screen parent) {
+	public default boolean onDoubleClick(float mouseX, float mouseY, int mouseButton, @Nullable Screen parent) {
 		return this.onClick(mouseX, mouseY, mouseButton, parent);
 	}
 	
@@ -134,7 +134,7 @@ public interface IWidget {
 	 * 
 	 * @return a boolean indicating whether or not this event should propagate to widgets with lower priorities
 	 */
-	public default boolean onParentDoubleClick(int mouseX, int mouseY, int mouseButton, @Nullable Screen parent) {
+	public default boolean onParentDoubleClick(float mouseX, float mouseY, int mouseButton, @Nullable Screen parent) {
 		return this.onParentClick(mouseX, mouseY, mouseButton, parent);
 	}
 	
@@ -146,7 +146,7 @@ public interface IWidget {
 	 * @param mouseButton
 	 * @param parent screen
 	 */
-	public default void onMouseDragged(int mouseX, int mouseY, int dX, int dY, int mouseButton, @Nullable Screen parent) {}
+	public default void onMouseDragged(float mouseX, float mouseY, float dX, float dY, int mouseButton, @Nullable Screen parent) {}
 
 	/**
 	 * Called when the mouse is released over this widget
@@ -158,7 +158,7 @@ public interface IWidget {
 	 * @param parent screen
 	 * 
 	 */
-	public default void onMouseReleased(int mouseX, int mouseY, int button, @Nullable IWidget draggedWidget) {}
+	public default void onMouseReleased(float mouseX, float mouseY, int button, @Nullable IWidget draggedWidget) {}
 	
 	/**
 	 * Called between the time the events are processed and the screen is drawn
@@ -184,7 +184,7 @@ public interface IWidget {
 	 * @param parent screen
 	 * @param amount
 	 */
-	public default boolean onMouseWheeled(int mouseX, int mouseY, int amount, @Nullable Screen parent) {
+	public default boolean onMouseWheeled(float mouseX, float mouseY, int amount, @Nullable Screen parent) {
 		return true;
 	}
